+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.3.2 (2019-09-18)
+- Added **SparkRepositoryAdapter[A, B]**. It will allow a **SparkRepository[A]** to write/read a data store of type
+ **B** by using an implicit **DatasetConverter[A, B]**
+- Added trait **Converter[A, B]** that handles the conversion between an object of type A and an object of type **B**
+- Added abstract class **DatasetConverter[A, B]** that extends a **Converter[Dataset[A], Dataset[B]]**
+- Added auto-correction for `SparkRepository.findby(conditions)` method when we filter by case class field name instead of column name
+
 ## 0.3.1 (2019-08-23)
 - Added sequential mode in class `Stage`. Use can turn in on by setting `parallel` to *true*.
 - Added external data flow description in pipeline description